deleted location component and moved logic to App.js. Added NASA sattelite image
