+++ src/components/NASA/NASA.css
@@ -1,0 +1,3 @@
+.satImg {
+    width: 10vw;
+}
